+++ src/functions/project/build-artefact-ts-serverless-functions.spec.ts
@@ -86,11 +86,14 @@ describe('ProjectBuildArtefactForTypeScriptServerlessFunctions', () => {
     await writeFile(resolve(tmpDir, '.npmrc'), '🔧');
     await writeFile(resolve(tmpDir, 'nope.js'), '🙈');
 
-    const actualArtefact = await context.call(ProjectBuildArtefact, {});
+    const actualArtefact = await context.call(ProjectBuildArtefact, {
+      artefact: 'my-archive.zip',
+    });
 
     expect(npmService.build).toHaveBeenCalledExactlyOnceWith({
       workingDirectory: tmpDir,
     });
+    expect(actualArtefact).toEqual(`${tmpDir}/my-archive.zip`);
     expect(await readArchiveContent(actualArtefact)).toEqual({
       'package.json': '{}',
       'package-lock.json': '🔒',